feat: manually request api routes on build (#142)
Adds an optional script for pulling a list of API routes to
`src/apiRoutes.json` on build.

did this at 3:30 in the morning so if it's cooked let me know

edit: one thing that I know isn't great is the documentation. there's
already so much (unnecessary?) pre-existing stuff that I'd like an
outside opinion on where to put docs for this script, if any more are
needed.

Fixes #140