feat: implement change password functionality  (#249)
This implements the functionality that allows a user to change their
password, both as an unauthenticated ("forgot password" in login screen)
and an authenticated user (navigate to the "reset password" page as an
authenticated user and change your password). Specifically, it allows
the user flow detailed in #194 to occur. Appropriate tests, types,
helper functions, and components are all added.

This also closes #146. Modifications were made to the specifications of
#146 to appropriately accommodate this.

Closes #194
Closes #146.