refactor: Forward requests on non protected authenticated users (#277)
## Checklist

<!-- Thank you for submitting a pull request to our repo. -->

<!-- If this is your first PR in this repo, please run through the
checklist
below to ensure a smooth review and merge process for your PR. -->

- [X] You've read the [Contributor
Guide](https://github.com/toddagriscience/Nightcrawler/blob/main/.github/CONTRIBUTING.md).
- [ ] You've included unit or integration tests for your change, where
applicable.
- [X] You've included inline docs for your change, where applicable.
- [X] There's an open issue for the PR that you are making. If you'd
like to propose a new feature or change, please open an issue to discuss
the change or find an existing issue.

<!-- Once all that is done, you're ready to go. Open the PR with the
content below. -->

## Description
From what I'm gathering, the issue described in #250 was caused by the
auth middleware intercepting requests and auto redirecting users back to
`/`. This change uses `NextResponse.next()` to allow the request to
"passthrough" as normal so it can be intercepted later on by the router
and redirected to the 404 like any other route.

I'm unable to test this at the moment so if someone could that would be
great.

Fixes #250

---------

Co-authored-by: kurealnum <oscar.gaske.cs@gmail.com>
Co-authored-by: Oscar <116363763+kurealnum@users.noreply.github.com>