refactor: homepage (#306)
Closes #264. This is a bit of a rough draft, feel free to make
critiques. This is a draft PR because I don't think we should merge this
yet, even with suggestions implemented. Here's a few things that we
should probably do first:

- Integrate thumbnails that complement the colors on the landing page
(they need to work with both dark and light mode).
- Add a few of the accent/highlight colors from the Figma. I know we're
going for a monotone look, but I feel like this is way too bland as is.